@kspdeveloper Large Update, King is now correctly moving
- Optimisation changes
Instead of using multiple Lists of Vector 3s, the Scripts now use a Dictionary with the position of a piece as a key and the gameobject as the value, this makes coding much easier and efficient.

- Now track protected pieces:
This is solely for the king, so when a king looks at it's possible moves, it also checks if a piece that is takeable is protected by another piece making it untakeable.

- King now correctly cannot move into spaces where it can be took. The main issue arrising from this was the pawns, where they can take diagnally, however the original way this was being calculated was looking at all possible moves but a diagnal take is not a possible move unless there is a piece currenlty on that square, this required a seperate funciton to check for the issue. and has now been accounted for.

- UI is now updated every round as opposed to on Update.