@kspdeveloper Castling, Forsyth–Edwards Notation, En Passant all fully working!
Castling:
Both White and Black Kings can now castle, they are also prevented from castling when
1. And / Either the rook or king is not in it's starting Position
2. When the King crosses the line of an enemy piece to perfrom a castle.
:
Forsyth–Edwards Notation (FEN):
This was required to translate the board to a FEN String so stockfish can read it, this has now been fully tested and implemented (Stockfish is not yet reading in the FEN state though). All 6 fields are correctly filled.

En Passant:
A rule which when the Pawn jumps two moves it is takeable by a pawn adjacent to it.
This has been tested and is fully working and also feeds back into the FEN String.